+++ library/src/methods/index.ts
@@ -9,6 +9,7 @@ export * from './getFallback/index.ts';
 export * from './getFallbacks/index.ts';
 export * from './is/index.ts';
 export * from './keyof/index.ts';
+export * from './message/index.ts';
 export * from './omit/index.ts';
 export * from './parse/index.ts';
 export * from './parser/index.ts';