+++ library/src/methods/message/message.test-d.ts
@@ -5,11 +5,11 @@ import { message } from './message.ts';
 describe('message', () => {
   test('should return schema object', () => {
     const schema = string();
-    expectTypeOf(message(schema, '')).toEqualTypeOf<typeof schema>();
+    expectTypeOf(message(schema, 'message')).toEqualTypeOf<typeof schema>();
   });
 
   test('should return async schema object', () => {
     const schema = objectAsync({ key: string() });
-    expectTypeOf(message(schema, '')).toEqualTypeOf<typeof schema>();
+    expectTypeOf(message(schema, 'message')).toEqualTypeOf<typeof schema>();
   });
 });